Instantiates ExecutionPlanCaptureCallbackImpl using ShimLoader
