+++ sql-plugin/src/main/scala/com/nvidia/spark/rapids/ShimLoader.scala
@@ -35,7 +35,7 @@ import org.apache.spark.sql.Strategy
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 import org.apache.spark.sql.catalyst.rules.Rule
 import org.apache.spark.sql.execution.{ColumnarRule, SparkPlan}
-import org.apache.spark.sql.rapids.AdaptiveSparkPlanHelperShim
+import org.apache.spark.sql.rapids.{AdaptiveSparkPlanHelperShim, ExecutionPlanCaptureCallbackBase}
 import org.apache.spark.sql.rapids.execution.UnshimmedTrampolineUtil
 import org.apache.spark.util.MutableURLClassLoader
 
@@ -421,4 +421,9 @@ object ShimLoader extends Logging {
     ShimReflectionUtils.newInstanceOf[AdaptiveSparkPlanHelperShim](
       "com.nvidia.spark.rapids.AdaptiveSparkPlanHelperImpl"
     )
+
+  def newExecutionPlanCaptureCallbackBase(): ExecutionPlanCaptureCallbackBase =
+    ShimReflectionUtils.
+        newInstanceOf[ExecutionPlanCaptureCallbackBase](
+          "org.apache.spark.sql.rapids.ExecutionPlanCaptureCallbackImpl")
 }
